% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{geometric_set_operations}
\alias{geometric_set_operations}
\alias{st_parallel_difference}
\alias{st_parallel_intersection}
\alias{st_parallel_sym_difference}
\title{Geometric operations on pairs of simple feature geometry sets (parallelized)}
\usage{
st_parallel_difference(x, y, threads = 1)

st_parallel_intersection(x, y, threads = 1)

st_parallel_sym_difference(x, y, threads = 1)
}
\arguments{
\item{x}{object of class \code{sfc} or \code{sf}.}

\item{y}{object of class \code{sfc} or \code{sf}.}

\item{threads}{\code{integer} number of threads for processing data. Defaults
to 1.}
}
\value{
The intersection, difference or symmetric difference between two
  sets of geometries.
}
\description{
Perform geometric set operations with simple feature
collections using parallel processing.
}
\details{
These functions are essentially just a wrapper for
  \code{\link[sf]{st_difference}}, \code{\link[sf]{st_intersection}},
  \code{\link[sf]{st_sym_difference}}.
}
\examples{
# create x object
pl1 <- sf::st_polygon(list(rbind(c(-1, -1), c(1, -1), c(1, 1), c(-1, 1),
                                 c(-1, -1))))
 pl2 <- pl1 + 2
 pl3 <- pl1 + c(-0.2, 2)
 x <- sf::st_sf(v = letters[1:3], geometry = sf::st_sfc(pl1, pl2, pl3),
                                                        crs = 3395)

# create z object
 pl4 <- pl1 * 0.8
 pl5 <- pl4 * 0.5 + c(2, 0.7)
 pl6 <- pl4 + 1
 pl7 <- pl1 * 0.5 + c(2, -0.5)
 z <-  sf::st_sf(v = letters[4:7], geometry = sf::st_sfc(pl4, pl5, pl6, pl7),
                 crs = 3395)

# calculate difference using sf::st_difference
y1 <- sf::st_difference(x, z)

# calculate difference using parallelized implementation
y2 <- st_parallel_difference(x, z, threads = 2)

# calculate intersection using sf::st_intersection
y3 <- sf::st_intersection(x, z)

# calculate intersection using parallelized implementation
y4 <- st_parallel_intersection(x, z, threads = 2)

# calculate symmetric difference using sf::st_sym_difference
y5 <- sf::st_sym_difference(x, z)

# calculate symmetric difference using parallelized implementation
y6 <- st_parallel_sym_difference(x, z, threads = 2)

# plot objects for visual comparison
par(mfrow = c(4, 2))
plot(st_geometry(x), main = "x", axes = TRUE, col = "lightblue")
plot(st_geometry(z), main = "z", axes = TRUE, col = "lightblue")
plot(st_geometry(y1), main = "sf::st_difference", axes = TRUE,
     col = "lightblue")
plot(st_geometry(y2), main = "st_parallel_difference", axes = TRUE,
     col = "lightblue")
plot(st_geometry(y3), main = "sf::st_intersection", axes = TRUE,
     col = "lightblue")
plot(st_geometry(y4), main = "st_parallel_intersection", axes = TRUE,
     col = "lightblue")
plot(st_geometry(y5), main = "sf::st_sym_difference", axes = TRUE,
     col = "lightblue")
plot(st_geometry(y6), main = "st_parallel_sym_difference", axes = TRUE,
     col = "lightblue")
}
\seealso{
\code{\link[sf]{st_difference}}, \code{\link[sf]{st_intersection}},
  \code{\link[sf]{st_sym_difference}}.
}
