% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga.R
\name{saga_difference}
\alias{saga_difference}
\title{SAGA Spatial Difference}
\usage{
saga_difference(x, y, saga_env = RSAGA::rsaga.env())
}
\arguments{
\item{x}{\code{\link[sf]{sfc}} object.}

\item{y}{\code{\link[sf]{sfc}} object.}

\item{saga_env}{\code{\link[base]{list}} object that contains information
for interfacing with SAGA. This object is typically created using
\code{\link[RSAGA]{rsaga.env}}.}
}
\value{
\code{\link[sf]{sf}} object.
}
\description{
This function performs a spatial difference using the System for Automated
Geoscientific Analyses (SAGA) software.
}
\details{
This function requires the SAGA software to be installed on the
  system. For information on installing SAGA, refer to:
  https://sourceforge.net/p/saga-gis/wiki/Binary%20Packages/.
}
\examples{
# load sf package
librarys(sf)

# make example data
b0 <- sf::st_polygon(list(rbind(c(-1, -1), c(1, -1), c(1, 1), c(-1, 1),
                                c(-1, -1))))
b1 <- b0 + 2
b2 <- b0 + c(-0.2, 2)
x <- st_union(st_sfc(b0, b1, b2))
a0 <- b0 * 0.8
a1 <- a0 * 0.5 + c(2, 0.7)
a2 <- a0 + 1
a3 <- b0 * 0.5 + c(2, -0.5)
y <- st_union(st_sfc(a0, a1, a2, a3))

# make data for comparison
o1 <- st_difference(x, y)
o2 <- saga_difference(x, y)

# test if sf and saga yield same result
st_precision(o1) <- 1000
st_precision(o2) <- 1000
st_equals(o1, o2, sparse = FALSE)[1] # TRUE

# plot sf and saga outputs for visual comparison
par(mfrow = c(1, 2))
plot(o1, main = "st_difference")
plot(o2, main = "saga_difference")
}
\seealso{
\code{\link[sf]{st_difference}}, \url{http://www.saga-gis.org}.
}
