% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saga.R
\name{saga_union}
\alias{saga_union}
\title{SAGA Spatial Union}
\usage{
saga_union(x, saga_env = RSAGA::rsaga.env())
}
\arguments{
\item{x}{\code{\link[sf]{sfc}} object.}

\item{saga_env}{\code{\link[base]{list}} object that contains information
for interfacing with SAGA. This object is typically created using
\code{\link[RSAGA]{rsaga.env}}.}
}
\value{
\code{\link[sf]{sf}} object.
}
\description{
This function performs a spatial union using the System for Automated
Geoscientific Analyses (SAGA) software.
}
\details{
This function requires the SAGA software to be installed on the
  system. For information on installing SAGA, refer to:
  https://sourceforge.net/p/saga-gis/wiki/Binary%20Packages/.
}
\seealso{
\code{\link[sf]{st_union}}, \url{http://www.saga-gis.org}.
}
