% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{st_parallel_make_valid}
\alias{st_parallel_make_valid}
\title{Make an invalid geometry valid (parallelized)}
\usage{
st_parallel_make_valid(x, threads = 1)
}
\arguments{
\item{x}{object of class \code{sfc}, \code{sfg} or \code{sf}.}

\item{threads}{\code{integer} number of cores for processing data.}
}
\value{
Object with fixed geometry.
}
\description{
Fix geometry issues using parallelized computations.
}
\details{
This function is essentially just a wrapper for
  \code{\link[lwgeom]{st_make_valid}}. See the documentation for
  \code{\link[lwgeom]{st_make_valid}} for more information.
}
\seealso{
\code{\link[lwgeom]{st_make_valid}}.
}
