% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{st_parallel_simplify}
\alias{st_parallel_simplify}
\title{Simplify geometry (parallelized)}
\usage{
st_parallel_simplify(x, preserveTopology = FALSE, dTolerance = 0,
  threads = 1)
}
\arguments{
\item{x}{Object of class \code{sfc} or \code{sf}.}

\item{preserveTopology}{\code{logical} carry out topology preserving
simplification? Defaults to \code{FALSE}.}

\item{dTolerance}{\code{numeric} tolerance parameter affecting how much
the geometry is simplified.}

\item{threads}{\code{integer} number of threads for processing data. Defaults
to 1.}
}
\value{
Simplified object.
}
\description{
Simplify the geometry in a simple feature data set using parallel processing.
}
\details{
This function is a wrapper for \code{\link[sf]{st_simplify}}.
}
\examples{
# load data
nc <- sf::read_sf(system.file("shape/nc.shp", package = "sf")) \%>\%
      sf::st_transform(3395)

# simplify data to 1 km
nc_simplified <- sf::st_simplify(nc, dTolerance = 5000)

# plot data for visual comparisons
par(mfrow = c(1, 2))
plot(st_geometry(nc), col = "white")
plot(st_geometry(nc_simplified), col = "white")
}
\seealso{
\code{\link[sf]{st_simplify}}.
}
