% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{st_parallel_transform}
\alias{st_parallel_transform}
\title{Transform or convert coordinates of simple feature (parallelized)}
\usage{
st_parallel_transform(x, crs, ..., threads = 1)
}
\arguments{
\item{x}{object of class \code{sfc}, \code{sfg} or \code{sf}.}

\item{crs}{coordinate reference system: integer with the EPSG code, or
character with proj4string.}

\item{...}{not used.}

\item{threads}{\code{integer} number of cores for processing data.}
}
\value{
Object with reprojected coordinates.
}
\description{
Reproject spatial data using parallelized computations.
}
\details{
This function is essentially just a wrapper for
  \code{\link[sf]{st_transform}}. See the documentation for
  \code{\link[sf]{st_transform}} for more information.
}
\seealso{
\code{\link[sf]{st_transform}}.
}
