% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{st_parallel_union}
\alias{st_parallel_union}
\title{Combine or union feature geometries (parallelized)}
\usage{
st_parallel_union(x, threads = 1)
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{threads}{\code{integer} number of threads for processing data. Defaults
to 1.}
}
\value{
A single geometry \code{sfc} object with resolved boundaries.
}
\description{
Combine several feature geometries into one, with or without
resolving internal boundaries
}
\details{
This function is essentially just a wrapper for
  \code{\link[sf]{st_union}}. See the documentation for
  \code{\link[sf]{st_union}} for more information.
}
\examples{
# load data
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"))

# create unioned data
ncu <- st_parallel_union(nc, threads = 1)

# plot data for visual comparison
par(mfrow = c(1, 2))
plot(sf::st_geometry(nc), main = "original data", col = "white")
plot(ncu, main = "unioned data", col = "white")
}
