% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{st_remove_holes}
\alias{st_remove_holes}
\title{Removes holes}
\usage{
st_remove_holes(x)
}
\arguments{
\item{x}{object of class \code{sfc}, \code{sfg} or \code{sf}.}
}
\value{
Object with holes removed.
}
\description{
Remove holes from polygons or multipolygons.
}
