% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_clean.R
\name{wdpa_clean}
\alias{wdpa_clean}
\title{Clean data}
\usage{
wdpa_clean(x, crs = sf::st_crs(3395), threads = 1)
}
\arguments{
\item{x}{\code{\link[sf]{sf}} object containing data from the WDPA.}

\item{crs}{\code{character} coordinate reference system in PROJ4 format.
Defaults to \code{sf::st_crs(3395)} (Mercator).}

\item{threads}{\code{numeric} number of threads to use for processing.
Defaults to 1.}
}
\value{
\code{\link[sf]{sf}} object.
}
\description{
Clean data from the World Database on Protected Areas (WDPA).
}
\details{
This function cleans the data by performing the
}
\examples{

}
\references{

}
\seealso{
\code{\link{wdpa_fetch}}.
}
