% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_fetch.R
\name{wdpa_fetch}
\alias{wdpa_fetch}
\title{Fetch data from the World Database on Protected Areas}
\usage{
wdpa_fetch(x, download_dir = tempdir(), force_download = FALSE,
  verbose = interactive())
}
\arguments{
\item{x}{\code{character} Country for which to download data. This argument
can be the name of the country (e.g. \code{"Liechtenstein"}) or the
ISO-3 code for the country (e.g. \code{"LIE"}). This argument can also
be set to \code{"global"} to download all of the protected areas available
in the database (approx. 1.1 GB).}

\item{download_dir}{\code{character} folder path to download the data.
Defaults to the temporary directory (\code{\link{tempdir}}).}

\item{force_download}{\code{logical} if the data has previously been
downloaded and is available at argument to \code{x}, should the data be
redownloaded anyway? Defaults to \code{FALSE}.}

\item{verbose}{\code{logical} should a progress on downloading data be
reported? Defaults to \code{TRUE} if the session is interactive.}
}
\value{
\code{\link[sf]{sf}} simple features object.
}
\description{
Download data from the World Database on Protected Areas (WDPA)
(available at \url{http://protectedplanet.net}) and import it. Note that
if the data has previously been downloaded and is available
in the specified download folder, then the data will simply be imported.
}
\details{
This function will simply download the specified protected area
  data and return it for subsequent use. It is strongly recomended that users
  clean the data prior to analysis. It is worth noting that the returned data
  for protected areas represented using both polygon and point geometries.
  Check out the \code{\link{clean_wdpa}} function to clean the data
  according to standard practices.
}
\examples{
\donttest{
# fetch data for Liechtenstein
lie_data <- wdpa_fetch("Liechtenstein")

# fetch data for Liechtenstein using the ISO3 code
lie_data <- wdpa_fetch("LIE", force_download = TRUE)
}
\dontrun{
# fetch data for the globe
# note that this might take some time given that the global data set is
# over 1 GB in size
global_data <- wdpa_fetch("global")
}
}
\seealso{
\code{\link{wdpa_clean}}, \code{\link[countrycode]{countrycode}},
  \url{http://protectedplanet.net}
}
