% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_latest_version.R
\name{wdpa_latest_version}
\alias{wdpa_latest_version}
\title{Query latest version of the World Database on Protected Areas}
\usage{
wdpa_latest_version()
}
\value{
\code{character} version of the dataset.
}
\description{
Find the latest version of the World Database on Protected Areas dataset.
This is a character identifier representing the month and year (e.g.
\code{Sep2020}) the data were released.
}
\details{
The version number can be determined based on the file name for
the global dataset (currently available at:
\url{http://wcmc.io/wdpa_current_release})
}
\examples{
\dontrun{
# find the latest version
wdpa_latest_version()
}

}
