% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_read.R
\name{wdpa_read}
\alias{wdpa_read}
\title{Read data from the World Database on Protected Areas}
\usage{
wdpa_read(x)
}
\arguments{
\item{x}{\code{character} file name for the zip archive downloaded from
\url{http://protectedplanet.net}.}
}
\value{
\code{\link[sf]{sf}} object.
}
\description{
Read data from the World Database on Protected Areas from a local file.
This function assumes that the data has already been downloaded to
your computer, see the \code{\link{wdpa_fetch}} function for automatically
downloading and importing the data into the current session.
}
\examples{
\donttest{
# find working url for Liechtenstein data set
last_few_months <- format(Sys.Date() - (seq(0, 10) * 30), "\%b\%Y")
potential_urls <- paste0("https://www.protectedplanet.net/downloads/WDPA_",
                         last_few_months, "_LIE?type=shapefile")
found_url <- FALSE
for (i in seq_along(potential_urls)) {
  if (!httr::http_error(potential_urls[i])) {
    download_url <- potential_urls[i]
    found_url <- TRUE
    break()
  }
}

# download the data set if working url was found
if (found_url) {
  # path to save file zipfile with data
  path <- tempfile(fileext = ".zip")

  # download zipfile
  result <- httr::GET(download_url, httr::write_disk(path))

  # load data
  lie_data <- wdpa_read(path)

  # plot data
  plot(lie_data)
}
}
}
\seealso{
\code{\link{wdpa_fetch}}, \code{\link{wdpa_clean}},
  \url{http://protectedplanet.net}.
}
