% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_stage_data.R
\name{wdpa_stage_data}
\alias{wdpa_stage_data}
\title{Stage data for download from the World Database on Protected Areas}
\usage{
wdpa_stage_data(x, wait = TRUE)
}
\arguments{
\item{\code{character}}{country for which to download data. This argument
can be the name of the country (e.g. \code{"Liechtenstein"}) or the
ISO-3 code for the country (e.g. \code{"LIE"}).}

\item{\code{logical}}{wait until the data is finished staging? Defaults to
\code{TRUE}.}
}
\value{
\code{logical} indicating success.
}
\description{
The World Database on Protected Areas updates rather frequently (at least
once a month). After the database is updated, the country-level data is
not immediately available for download. This function directs
the \href{http://protectedplanet.net}{ProtectedPlanet} website to
start preparing the data for later download. Once the data for a country
has been staged for downloading, this step does not need to be repated
until the database is updated again.
}
\details{
Note that the global data set does not need to be staged and is
  immediately available for download after the database has been updated.
}
\examples{
\donttest{
wdpa_stage_data("LIE")
}
}
\seealso{
\code{\link{wdpa_url}}.
}
