% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_url.R
\name{wdpa_url}
\alias{wdpa_url}
\title{Download URL for the World Database on Protected Areas}
\usage{
wdpa_url(x)
}
\arguments{
\item{x}{\code{character} Country for desired data. This argument
can be the name of the country (e.g. \code{"Liechtenstein"}) or the
ISO-3 code for the country (e.g. \code{"LIE"}). This argument can also
be set to \code{"global"} to obtain the URL for the global data set.}
}
\value{
\code{character} URL.
}
\description{
Obtain the URL for downloading data from the World Database on Protected
Areas (WDPA).
}
\examples{
\donttest{
# obtain url for USA data
usa_url <- wdpa_url("United States of America")
print(usa_url)

# obtain url for USA data (using the ISO3 code)
usa_url <- wdpa_url("USA")
print(usa_url)

# download link for global data
global_url <- wdpa_url("global")
print(global_url)
}
}
\seealso{
\code{\link{wdpa_fetch}}, \code{\link[countrycode]{countrycode}}.
}
