% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{wdpar}
\alias{wdpar}
\alias{wdpar-package}
\title{wdpar: Interface to the World Database on Protected Areas}
\description{
The \emph{wdpar R} package provides an interface to the World Database on
Protected Areas (WDPA). It provides functions for automatically downloading
data (from \href{http://protectedplanet.net}{Protected Planet}) and cleaning
data following best practices (outlined in
\href{https://dx.doi.org/10.1111/conl.12158}{Butchart \emph{et al.} 2015};
\href{https://dx.doi.org/10.1126/science.aac9180}{Runge \emph{et al.} 2015}.
The main functions are \code{\link{wdpa_fetch}} for downloading data and
\code{\link{wdpa_clean}} for cleaning data. For more information, please see
the package vignette.
}
