% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{geometric_set_operations}
\alias{geometric_set_operations}
\alias{st_parallel_difference}
\alias{st_parallel_intersection}
\alias{st_parallel_sym_difference}
\title{Geometric operations on pairs of simple feature geometry sets (parallelized)}
\usage{
st_parallel_difference(x, y, threads = 1)

st_parallel_intersection(x, y, threads = 1)

st_parallel_sym_difference(x, y, threads = 1)
}
\arguments{
\item{x}{object of class \code{sfc} or \code{sf}.}

\item{y}{object of class \code{sfc} or \code{sf}.}

\item{threads}{\code{integer} number of cores for processing data.}
}
\value{
The intersection, difference or symmetric difference between two
  sets of geometries.
}
\description{
Perform geometric set operations with simple feature
collections using parallel processing.
}
\details{
These functions are essentially just a wrapper for
  \code{\link[sf]{st_difference}}, \code{\link[sf]{st_intersection}},
  \code{\link[sf]{st_sym_difference}}.
}
\seealso{
\code{\link[sf]{st_difference}}, \code{\link[sf]{st_intersection}},
  \code{\link[sf]{st_sym_difference}}.
}
