% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_and_eez_fetch.R
\name{land_and_eez_fetch}
\alias{land_and_eez_fetch}
\title{Fetch land and exclusive economic zone data}
\usage{
land_and_eez_fetch(x, crs = 3395, download_dir = tempdir(), threads = 1,
  verbose = interactive())
}
\arguments{
\item{x}{\code{character} Country for which to download data. This argument
can be the name of the country (e.g. \code{"Liechtenstein"}) or the
ISO-3 code for the country (e.g. \code{"LIE"}).}

\item{crs}{\code{character} coordinate reference system in PROJ4 format.
Defaults to \code{3395} (Mercator).}

\item{download_dir}{\code{character} folder path to download the data.
Defaults to the temporary directory (\code{\link[base]{tempdir}}).}

\item{threads}{\code{numeric} number of threads to use for processing.
Defaults to 1.}

\item{verbose}{\code{logical} should download progress be reported? Defaults
to \code{TRUE} if executed in an interactive session.}
}
\value{
\code{\link[sf]{sf}} spatial data object. This object contains the
 following two fields: \code{"ISO3"} which shows the country associated
 with each geometry, \code{"TYPE"} which indicates if a geometry corresponds
 to the country's terrestrial land mass (\code{"LAND"}) or its
 exclusive economic zone (\code{"EEZ"}).
}
\description{
Fetch and assemble land and exclusive economic zone (EEZ) data for the
world's nations.
}
\details{
Data are assembled using the following steps.
\enumerate{
\item Spatial data showing the spatial extent of each country's terrestrial
  administrative area is downloaded (from the Database of Global
  Administrative Areas; \url{http://www.gadm.org/})
\item Spatial data showing the spatial extent of each country's exclusive
  economic zone are downloaded (from \url{http://marineregions.org}; approx.
  113 MB in size; Claus \emph{at al.} 2017))
\item Both data sets are scanned for invalid geometries, and invalid
  geometries are fixed (using \code{\link{st_parallel_make_valid}}).
\item Both data sets are reprojected to argument to \code{crs}.
\item Scan both data sets again for invalid geometries, and fix any
  invalid geometries that have manifested (using
  \code{\link{st_parallel_make_valid}}).
\item A field denoting the country code is created in the two data sets
  is created (\code{"ISO3"}).
\item Both data sets are dissolved by the country code field (\code{"ISO3"})
  to remove overlapping geometries.
\item Scan both data sets again for invalid geometries, and fix any
  invalid geometries that have manifested (using
  \code{\link{st_parallel_make_valid}}).
\item Areas in the exclusive economic zone datset that overlap with the
  terrestrial data set are removed (using \code{\link[sf]{st_difference}}).
\item The terrestrial data is assigned a new field \code{"TYPE"} containing
  \code{"LAND"}.
\item The exclusive economic zone data is assigned a new field
  \code{"TYPE"} containing \code{"EEZ"}.
\item The terrestrial and exclusive economic zone datsets are merged
  (using \code{\link[base]{rbind}}).
\item The merged data set is sorted by country code (\code{"ISO3"}) and
  (\code{"TYPE"}; using \code{\link[sf]{filter.sf}}).
}
}
\examples{
\donttest{
# fetch land and eez data for New Zealand
nzl_data <- land_and_eez_fetch("NZL")
}
\dontrun{
# fetch land and eez data for the planet
global_data <- land_and_eez_fetch("globe")
}
}
\references{
Claus S, De Hauwere N, Vanhoorne B, Souza Dias F, Oset Garc\'{i}a P,
Hernandez F & Mees J (2017) MarineRegions.org. Flanders Marine Institute.
Available at \url{http://www.marineregions.org}.
}
