% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{st_parallel_transform}
\alias{st_parallel_transform}
\title{Transform or convert coordinates of simple feature (parallelized)}
\usage{
st_parallel_transform(x, crs, ..., threads = 1)
}
\arguments{
\item{x}{object of class \code{sfc} or \code{sf}.}

\item{crs}{coordinate reference system: integer with the EPSG code, or
character with proj4string.}

\item{...}{not used.}

\item{threads}{\code{integer} number of threads for processing data. Defaults
to 1.}
}
\value{
Object with reprojected coordinates.
}
\description{
Reproject spatial data using parallelized computations.
}
\details{
This function is essentially just a wrapper for
  \code{\link[sf]{st_transform}}. See the documentation for
  \code{\link[sf]{st_transform}} for more information.
}
\examples{
# create data
pl1 <- sf::st_polygon(list(matrix(c(0, 0, 2, 0, 2, 2, 0, 2, 0, 0),
                                   byrow = TRUE, ncol = 2)))
pl2 <- sf::st_polygon(list(matrix(c(5, 5, 7, 5, 7, 7, 5, 7, 5, 5),
                                   byrow = TRUE, ncol = 2)))
x <- sf::st_sf(z = runif(2), geometry = sf::st_sfc(list(pl1, pl2),
                                                   crs = 4326))

# transform data using sf::st_transform
y1 <- sf::st_transform(x, 3395)

# transform data using parallelized implementation
y2 <- st_parallel_transform(x, 3395, threads = 2)

# plot objects for visual comparison
par(mfrow = c(1, 2))
plot(y1, main = "sf::st_transform", axes = TRUE)
plot(y2, main = "st_parallel_transform", axes = TRUE)
}
\seealso{
\code{\link[sf]{st_transform}}.
}
