% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_clean.R
\name{wdpa_clean}
\alias{wdpa_clean}
\title{Clean data from the World Database on Protected Areas}
\usage{
wdpa_clean(x, crs = 3395, tolerance = 1, threads = 1, verbose = FALSE)
}
\arguments{
\item{x}{\code{\link[sf]{sf}} object containing data from the WDPA.}

\item{crs}{\code{character} coordinate reference system in PROJ4 format.
Defaults to \code{3395} (Mercator).}

\item{tolerance}{\code{numeric} tolerance for snapping geometry to a
grid for resolving invalid geometries.}

\item{threads}{\code{numeric} number of threads to use for processing.
Defaults to 1.}

\item{verbose}{\code{logical} should progress on data cleaning be reported?
Defaults to \code{FALSE}.}
}
\value{
\code{\link[sf]{sf}} object.
}
\description{
Clean data from the World Database on Protected Areas (WDPA).
}
\details{
This function cleans data from World Database on Protected Areas
  following best practices (Butchart \emph{et al.} 2015, Runge \emph{et al.}
  2015,
  \url{https://protectedplanet.net/c/calculating-protected-area-coverage}).
  \enumerate{
  \item Repair invalid geometry (using \code{\link{st_parallel_make_valid}}).
  \item Exclude protected areas that are not currently implemented
    (retain areas with the status \code{"Designated"},
    \code{"Inscribed"}, \code{"Established"}).
  \item Exclude UNESCO Biosphere Reserves (Coetzer \emph{et al.} 2014).
  \item Create a field (\code{"GEOMTRY_TYPE"}) indicating if areas are
    represented as point localities (\code{"POINT"}) or as polygons
    (\code{"POLYGON"}).
  \item Exclude areas represented as point localities that do not
    have a reported spatial extent (i.e. missing data for the field
  \item Reproject data to coordinate system specified in argument to
    \code{crs} (using \code{\link{st_parallel_transform}}).
  \item Repair any invalid geometry that has since manifested (using
    \code{\link{st_parallel_make_valid}}).
  \item Buffer areas represented as point localities to circular areas
    using their reported spatial extent (using data in the field
    \code{"REP_AREA"} and \code{\link[sf]{st_buffer}}).
  \item Assemble land and exclusive economic zone data (using
    \code{\link{land_and_eez_fetch}}).
  \item Extract terrestrial protected areas and erase localities that occur
   in the ocean.
  \item Extract marine protected areas and erase regions that occur on land.
  \item Merge terrestrial and marine protected area data sets.
  \item The \code{"MARINE"} field is converted from integer codes
    to descriptive name (\code{0} = \code{"terrestrial"},
    \code{1} = \code{"partial"}, \code{2} = \code{"marine"}).
  \item Zeroes in the \code{"STATUS_YR"} field are replaced with \code{NA}
    values.
  \item Zoers in the \code{"NO_TK_AREA"} field are replaced with \code{NA}
    values for areas for with such data are not reported or applicable
    (i.e. areas with the values \code{"Not Applicable"}
    or \code{"Not Reported"} in the \code{"NO_TK_AREA"} field).
  \item Snap geometries to a grid to fix any unresolved geometry issues
    (using \code{link[lwgeom]{st_snap_to_grid}}).
  \item Repair any invalid geometry that has manifested (using
    \code{\link{st_parallel_make_valid}}).
  \item Erase overlapping geometries (discussed in Deguignet \emph{et al.}
    2017). Geometries are erased such that protected areas associated with
    more effective management categories (\code{"IUCN_CAT"}) or have
    historical precdence are retained (using
    \code{\link[sf]{st_difference}}).
  \item Slivers are removed (geometries an area less than 1e-10 square
    kilometers).
  \item Calculate size of areas in square kilometers (stored in field
    \code{"AREA_KM2"}).
 }
}
\examples{
\donttest{
# fetch data for the Marshall Islands
mhl_raw_data <- wdpa_fetch("MHL", wait = TRUE)

# clean data
mhl_data <- wdpa_clean(mhl_raw_data)

# plot cleaned data set
plot(mhl_data)

# plot geometries for visual comparison
par(mfrow = c(1, 2))
plot(sf::st_geometry(mhl_raw_data), main = "orginal data", col = "white")
plot(sf::st_geometry(mhl_data), main = "cleaned data", col = "white")

\dontrun{
# fetch data for all protected areas on the planet
# note that this might take some time given that the global data set is
# over 1 GB in size
global_raw_data <- wdpa_fetch("global")

# set number of threads for processing
n_threads <- max(1, parallel::detectCores(TRUE) - 1)

# clean global data set using parallel processing
global_data <- wdpa_clean(global_raw_data, threads = n_threads)

# plot data
plot(global_data)
}}
}
\references{
Butchart SH, Clarke M, Smith RJ, Sykes RE, Scharlemann JP,
Harfoot M, ... & Brooks TM (2015) Shortfalls and solutions for
meeting national and global conservation area targets.
\emph{Conservation Letters}, \strong{8}: 329--337.

Coetzer KL, Witkowski ET, & Erasmus BF (2014) Reviewing
Biosphere Reserves globally: Effective conservation action or bureaucratic
label? \emph{Biological Reviews}, \strong{89}: 82--104.

Deguignet M, Arnell A, Juffe-Bignoli D, Shi Y, Bingham H, MacSharry B &
Kingston N (2017) Measuring the extent of overlaps in protected area
designations. \emph{PloS One}, \strong{12}: e0188681.

Runge CA, Watson JEM, Butchart HM, Hanson JO, Possingham HP & Fuller RA
(2015) Protected areas and global conservation of migratory birds.
\emph{Science}, \strong{350}: 1255--1258.
}
\seealso{
\code{\link{wdpa_fetch}},
  \url{https://protectedplanet.net/c/calculating-protected-area-coverage}.
}
