% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_latest_version.R
\name{wdpa_latest_version}
\alias{wdpa_latest_version}
\title{Query latest version}
\usage{
wdpa_latest_version()
}
\value{
\code{character} version of the dataset.
}
\description{
Find the latest version of the combined
World Database on Protected Areas
(WDPA) and World Database on Other Effective Area-Based Conservation Measures
(WDOECM) dataset.
This is a character identifier representing the month and year (e.g.
\code{Sep2020}) the data were released.
}
\details{
The version number is determined using a web address where the
global dataset is available. For specific details, please refer to
the source code for this function.
}
\examples{
\dontrun{
# find the latest version
wdpa_latest_version()
}

}
