% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_read.R
\name{wdpa_read}
\alias{wdpa_read}
\title{Read data from the World Database on Protected Areas}
\usage{
wdpa_read(x)
}
\arguments{
\item{x}{\code{character} file name for the zip archive downloaded from
\url{http://protectedplanet.net}.}
}
\value{
\code{\link[sf]{sf}} object.
}
\description{
Read data from the World Database on Protected Areas from a local file.
This function assumes that the data has already been downloaded to
your computer, see the \code{\link{wdpa_fetch}} function for automatically
downloading and importing the data into the current session.
}
\examples{
# load in simulated WDPA dataset used for examples and testing
fake_data <- wdpa_read(system.file("inst/extdata/WDPA_Nov2017_FAKE.zip",
                                   package = "wdpar"))
# plot simulated WDPA dataset
plot(fake_data)
}
\seealso{
\code{\link{wdpa_fetch}}, \code{\link{wdpa_clean}},
  \url{http://protectedplanet.net}.
}
